+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/AttestationSubnetScorer.java
@@ -14,7 +14,7 @@
 package tech.pegasys.teku.networking.eth2.gossip;
 
 import java.util.function.IntUnaryOperator;
-import tech.pegasys.teku.networking.p2p.connection.PeerScorer;
+import tech.pegasys.teku.networking.eth2.peers.PeerScorer;
 import tech.pegasys.teku.networking.p2p.gossip.GossipNetwork;
 import tech.pegasys.teku.networking.p2p.peer.NodeId;
 import tech.pegasys.teku.ssz.SSZTypes.Bitvector;
