+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/peers/PeerScorer.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.networking.p2p.connection;
+package tech.pegasys.teku.networking.eth2.peers;
 
 import tech.pegasys.teku.networking.p2p.discovery.DiscoveryPeer;
 import tech.pegasys.teku.networking.p2p.peer.NodeId;
